dev: fix linter warns

+++ storage/s3.go
@@ -193,7 +193,7 @@ func (s *S3) upload(fileKey string) (err error) {
 			// set the part size as low as possible to avoid timeouts and aborts
 			// also set concurrency to 1 for the same reason
 			var partSize int64 = 5242880 // 5MiB
-			maxParts := math.Ceil(float64(info.Size() / partSize))
+			maxParts := math.Ceil(float64(info.Size()) / float64(partSize))
 
 			// 10000 parts is the limit for AWS S3. If the resulting number of parts would exceed that limit, increase the
 			// part size as much as needed but as little possible
